fix(punishment): use jedis.expire(String, long) instead of (String, int)
Newer jedis versions deprecate the int-seconds overload. The cast
to int was also a silent truncation hazard for TTLs near 2^31s.

+++ commons/src/main/java/net/swofty/commons/punishment/PunishmentRedis.java
@@ -88,7 +88,7 @@ public class PunishmentRedis {
             if (expiresAt > 0) {
                 long ttlSeconds = (expiresAt - System.currentTimeMillis()) / 1000;
                 if (ttlSeconds > 0) {
-                    jedis.expire(redisKey, (int) ttlSeconds);
+                    jedis.expire(redisKey, ttlSeconds);
                 }
             } else {
                 jedis.persist(redisKey);
